requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
requests==2.32.3
python-dotenv==1.0.1
pytesseract==0.3.13
Pillow==11.0.0
pdfplumber==0.11.4
python-docx==1.1.2
reportlab==4.2.5
langdetect==1.0.9
google-generativeai==0.8.3
